==> jegil32/cluster_baselineCan2025/02-3_preparation_donnees_sante.ipynb <==
# Préparation des variables de sante

library(tidyverse)
library(factoextra)

# Charger les données
DataPilot <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/01_pilot_2025.rds")

# Charger le fichier des variables
source("01-1_variables.R")

df_sante <- DataPilot %>%
  select(all_of(variables_sante))

# Boucle pour afficher la table pour chaque variable
for (v in variables_sante) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_sante[[v]]))
    cat("\n\n")
  }
}

### Transformations

- cons_meat : Devrait être une variable ordinale 
  - 0 Not more than once a month
  - 0.5 Few weekly
  - 1 Daily

- cons_..._Drink : On peut conserver tel quel

- cons_Smoke : Devrait être une variable ordinale
  - 0 Never
  - 0.5 Not daily
  - 1 Daily

# Transformation de cons_meat en variable ordinale
#-------------------------------------------------

df_sante <- df_sante %>%

  mutate(
    lifestyle_eatMeatFreq = if_else(
      lifestyle_eatMeatFreq == 0 | lifestyle_eatMeatFreq == 0.17 | lifestyle_eatMeatFreq == 0.33,
      0,
      if_else(
        lifestyle_eatMeatFreq == 0.5 | lifestyle_eatMeatFreq == 0.67,
        0.5,
        1
      )
    )
  ) 

# Transformation de cons_Smoke variable ordinale
#-------------------------------------------------

df_sante <- df_sante %>%

  mutate(
    lifestyle_smokeFreq = if_else(
      lifestyle_smokeFreq == 0,
      0,
      if_else(
        lifestyle_smokeFreq == 0.1667 | lifestyle_smokeFreq == 0.3333 | lifestyle_smokeFreq == 0.5 | lifestyle_smokeFreq == 0.6667,
        0.5,
        1
      )
    )
  )

variables_sante_clust <- c(
  "id",
  "lifestyle_eatMeatFreq",
  "lifestyle_favAlcoolRedWine",
  "lifestyle_favAlcoolWhiteWine",
  "lifestyle_favAlcoolRoseWine",
  "lifestyle_favAlcoolSpirits",
  "lifestyle_favAlcoolBubbleDrink",
  "lifestyle_favAlcoolBeer",
  "lifestyle_favAlcoolMicroBeer",
  "lifestyle_favAlcoolCocktail",
  "lifestyle_favAlcoolDontDrink",
  "lifestyle_smokeFreq"
)

# Boucle pour afficher la table pour chaque variable
for (v in variables_sante_clust) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_sante[[v]]))
    cat("\n\n")
  }
}

# Sauvegarder les données préparées
saveRDS(df_sante, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_sante.rds")

---

# SUGGESTION : Activité de Clustering

Si souhaité, effectuer un exercice de clustering avec ce sous-groupe de variables pour mieux saisir les données et leur relation. 

Cet exercice est purement exploratoire et sert à voir si les données peuvent être utilisées pour faire du clustering.
