==> jegil32/cluster_baselineCan2025/02-5_preparation_donnees_commerce.ipynb <==
# Préparation des variables commerce

library(tidyverse)
library(factoextra)

# Charger les données
DataPilot <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/01_pilot_2025.rds")

# Charger le fichier des variables
source("01-1_variables.R")

df_commerce <- DataPilot %>%
  select(all_of(variables_commerce))

# Boucle pour afficher la table pour chaque variable
for (v in variables_commerce) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_commerce[[v]]))
    cat("\n\n")
  }
}

### Transformations

- cons_brand : Aucun changement

- cons_coffee : Regrouper SC (second cup) avec Other

# Regroupement de cons_coffee
#-----------------------------

df_commerce <- df_commerce %>%

  mutate(
    lifestyle_consCoffeeOther = if_else(
      lifestyle_consCoffeeSecondCup == 1,
      1,
      lifestyle_consCoffeeOther
    )
  ) %>%
  select(-lifestyle_consCoffeeSecondCup)

variables_commerce_clust <- c(
  "id",
  "lifestyle_consClothesFrip",
  "lifestyle_consClothesIndependent",
  "lifestyle_consClothesChain",
  "lifestyle_consClothesSuperstores",
  "lifestyle_consClothesDepartment",
  "lifestyle_consClothesOnline",
  "lifestyle_consClothesOther",
  "lifestyle_consCoffeeTimHortons",
  "lifestyle_consCoffeeStarbucks",
  "lifestyle_consCoffeeMcDo",
  "lifestyle_consCoffeeOther",
  "lifestyle_consCoffeeIndependent",
  "lifestyle_consCoffeeNone"
)

# Boucle pour afficher la table pour chaque variable
for (v in variables_commerce_clust) {
  if (v != "id") {
    cat("Table pour la variable:", v, "\n")
    print(table(df_commerce[[v]]))
    cat("\n\n")
  }
}

# Sauvegarder les données préparées
saveRDS(df_commerce, file = "../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/02_pilot_2025_commerce.rds")

---

# SUGGESTION : Activité de Clustering

Si souhaité, effectuer un exercice de clustering avec ce sous-groupe de variables pour mieux saisir les données et leur relation. 

Cet exercice est purement exploratoire et sert à voir si les données peuvent être utilisées pour faire du clustering.
