==> jegil32/cluster_baselineCan2025/03-8_desc_complete.ipynb <==
# Library
library(ggplot2)
library(dplyr)

# Load Data and Clustering Results
DataPilot <- readRDS("../../_SharedFolder_datagotchi_federal_2024/data/pilote/dataClean/datagotchi2025_canada_pilotClustering_20250319.rds")
DataClustering <- readRDS("../../_PrivateFolder_datagotchi_federal_2025/data/clustering/can2025/03_pilot_2025.rds")
names(DataClustering)

# Join both datasets to have cluster, lifestyle and id

# Sélectionner les variables lifestyle et SES de dfLifestyle
DataPilot_selected <- DataPilot %>%
  select(id,
         lifestyle_exercise, 
         lifestyle_goFishingFreq_bin, 
         lifestyle_goHuntingFreq_bin, 
         lifestyle_goMuseumsFreq_bin, 
         lifestyle_performingArtsFreq_numeric,
         lifestyle_manualTaskFreq_numeric, 
         lifestyle_motorizedActFreq_bin, 
         lifestyle_unmotorizedActFreq_bin, 
         lifestyle_volunteeringFreq, 
         lifestyle_makeArtFreq_numeric,
         lifestyle_travelFreq_numeric,
         lifestyle_gamingFreq_numeric, 
         lifestyle_workField,
         lifestyle_workType,
         lifestyle_consClothes, 
         lifestyle_eatMeatFreq, 
         lifestyle_fridgeVegetalMilk, 
         lifestyle_fridgeTofuTempeh, 
         lifestyle_fridgeOrganicVeggies, 
         lifestyle_fridgeEnergyDrink,
         lifestyle_fridgePizza,
         lifestyle_fridgeButterTarts,
         lifestyle_fridgeBacon,
         lifestyle_fridgeCeasarDressing,
         lifestyle_fridgeHotSauce,
         lifestyle_fridgeSoftDrinks,
         lifestyle_fridgeGroundBeef,
         lifestyle_alcoolFreq,
         lifestyle_weedFreq,
         lifestyle_consCoffee, 
         lifestyle_ownPet, 
         lifestyle_hasTattoos, 
         lifestyle_mostFreqSocialMedia,
         lifestyle_socialMediaTime,
         lifestyle_typeMorningToEvening,
         lifestyle_typeTravel,
         dv_attitudeLeftvsRight, 
         dv_voteChoice,
         ses_age, 
         ses_religion_bin,
         ses_religiosity,
         ses_owner,
         issue_proQcIndependance,
         dv_voteChoice)

# Sélectionner les variables clusters de dfUsedForClustering
DataClustering_selected <- DataClustering %>%
  select(id, 
         cluster_name, 
         starts_with("cluster_"), 
         starts_with("rci"),
         dv_turnout,
         dv_solidity,
         dv_voteChoiceLPC,
         dv_voteChoiceCPC,
         dv_voteChoiceNDP,
         dv_voteChoiceBQ,
         dv_voteChoiceGPC,
         dv_voteChoiceOther,
         ses_genderMale,
         ses_languageEnglish,
         ses_languageFrench,
         ses_languageOther,
         ses_educBHS,
         ses_educPostHS,
         ses_educUniv,
         ses_incomeLow,
         ses_incomeMid,
         ses_incomeHigh,
         ses_immigrant,
         ses_ethnicityWhite,
         ses_ethnicityMinority,
         ses_sexOrientationHetero,
         ses_sexOrientationQueer,
         ses_regionPrairies,
         ses_regionBC,
         ses_regionAtlantic,
         ses_regionOntario,
         ses_regionQuebec,
         ses_regionTerritories,
         ses_regionWest,
         ses_urban,
         lifestyle_typeTransportCar,
         lifestyle_typeTransportNoCar,
         lifestyle_favAlcoolRedWine,
         lifestyle_favAlcoolWhiteWine,
         lifestyle_favAlcoolSpirits,
         lifestyle_favAlcoolBeer,
         lifestyle_favAlcoolCocktail,
         lifestyle_favAlcoolDontDrink,
         lifestyle_consCoffeeTimHortons,
         lifestyle_consCoffeeMcDo,
         lifestyle_consCoffeeStarbucks,
         lifestyle_consCoffeeIndependent,
         lifestyle_consCoffeeNone)



# Fusionner les deux jeux de données sur id
DataValidation <- DataClustering_selected %>%
  inner_join(DataPilot_selected, by = "id")

# Voir les variables de ce jeu fusionné
names(DataValidation)

DataValidation$cluster_name[DataValidation$cluster_name == 1] <- "Maxime"
DataValidation$cluster_name[DataValidation$cluster_name == 2] <- "Michel"
DataValidation$cluster_name[DataValidation$cluster_name == 3] <- "David"
DataValidation$cluster_name[DataValidation$cluster_name == 4] <- "Robert"
DataValidation$cluster_name[DataValidation$cluster_name == 5] <- "Emily"
DataValidation$cluster_name[DataValidation$cluster_name == 6] <- "Jennifer"
DataValidation$cluster_name[DataValidation$cluster_name == 7] <- "John"
DataValidation$cluster_name[DataValidation$cluster_name == 8] <- "Julie"
DataValidation$cluster_name[DataValidation$cluster_name == 9] <- "Arjun"
DataValidation$cluster_name[DataValidation$cluster_name == 10] <- "Zoe"
table(DataValidation$cluster_name)

library(dplyr)

# Sélectionner le cluster
cluster <- "John"
cluster_data <- DataValidation %>% filter(cluster_name == cluster)

# Définition d'une liste de variables avec leur étiquette de sortie
# Ici, on suppose que les variables binaires (0/1) servent à calculer des pourcentages,
# et que les variables continues (ex. âge) donnent une moyenne.
vars_to_print <- list(
  ses_genderMale = "Hommes",   # Variable binaire : 1 = homme, 0 = femme
  ses_age = "Age moyen",           # Variable numérique continue
  ses_immigrant = "Immigrants", # Variable binaire : 1 = immigrant, 0 = non-immigrant
  ses_languageFrench = "Francophones",
  ses_languageEnglish = "Anglophones",
  ses_languageOther = "Allophones",
  ses_educBHS = "Education pre-secondaire",
  ses_educPostHS = "Education secondaire",
  ses_educUniv = "Education universitaire",
  ses_incomeLow = "Revenu faible",
  ses_incomeMid = "Revenu moyen",
  ses_incomeHigh = "Revenu eleve",
  ses_immigrant = "Immigrants",
  ses_ethnicityMinority = "Minorites visibles",
  ses_sexOrientationQueer = "Queer (Non-hetero)",
  ses_regionPrairies = "Prairies",
  ses_regionBC = "Colombie-Britanique",
  ses_regionOntario = "Ontario",
  ses_regionQuebec = "Quebec",
  ses_regionAtlantic = "Provinces atlantiques",
  ses_regionTerritories = "Territoires",
  ses_urban             = "Urbain",
  ses_religion_bin            = "Religieux",
  ses_religiosity            = "Religiosite (sur 1)",
  ses_owner                  = "Proprietaire",
  lifestyle_exercise = "Exercice physique prefere",
  lifestyle_goFishingFreq_bin = "Pecheurs",
  lifestyle_goHuntingFreq_bin = "Chasseurs",
  lifestyle_goMuseumsFreq_bin = "Visite les musees",
  lifestyle_manualTaskFreq_numeric = "Frequence de travaux manuels (sur 1)",
  lifestyle_motorizedActFreq_bin = "Activites motorisees",
  lifestyle_unmotorizedActFreq_bin = "Activites non-motorisees",
  lifestyle_volunteeringFreq = "Frequence de benevolat (sur 1)",
  lifestyle_makeArtFreq_numeric = "Frequence de travaux artistiques (sur 1)",
  lifestyle_gamingFreq_numeric = "Frequence jeux videos (sur 1)",
  lifestyle_travelFreq_numeric = "Frequence de voyage (sur 1)",
  lifestyle_typeTravel           = "Type de voyage",
  lifestyle_typeTransportNoCar = "Pas de voiture",
  lifestyle_workField = "Domaine d'emploi",
  lifestyle_workType = "Type d'emploi",
  lifestyle_consClothes          = "Lieux d'achat des vetements",
  lifestyle_eatMeatFreq          = "Frequence de consommation de viande (sur 1)",
  lifestyle_fridgeVegetalMilk    = "Presence de lait vegetal au frigo",
  lifestyle_fridgeTofuTempeh     = "Presence de tofu/tempeh au frigo",
  lifestyle_fridgeOrganicVeggies = "Presence de legumes bio au frigo",
  lifestyle_fridgeEnergyDrink    = "Presence de boissons energisantes au frigo",
  lifestyle_fridgePizza          = "Presence de pizza au frigo",
  lifestyle_fridgeButterTarts    = "Presence de butter tarts au frigo",
  lifestyle_fridgeBacon          = "Presence de bacon au frigo",
  lifestyle_fridgeCeasarDressing = "Presence de sauce Cesar au frigo",
  lifestyle_fridgeHotSauce       = "Presence de sauce piquante au frigo",
  lifestyle_fridgeSoftDrinks     = "Presence de sodas au frigo",
  lifestyle_fridgeGroundBeef     = "Presence de viande hachee au frigo",
  lifestyle_alcoolFreq           = "Frequence de consommation d'alcool (sur 1)",
  lifestyle_weedFreq             = "Frequence de consommation de cannabis (sur 1)",
  lifestyle_consCoffeeTimHortons = "Tim",
  lifestyle_consCoffeeMcDo = "McDo",
  lifestyle_consCoffeeStarbucks = "Starbucks",
  lifestyle_consCoffeeIndependent = "independent",
  lifestyle_consCoffeeNone = "Aucun café",
  lifestyle_ownPet               = "Animaux de compagnie",
  lifestyle_favAlcoolRedWine     = "Vin rouge",
  lifestyle_favAlcoolWhiteWine   = "Vin blanc",
  lifestyle_favAlcoolSpirits     = "Spritis",
  lifestyle_favAlcoolBeer        = "Biere",
  lifestyle_favAlcoolCocktail   = "Cocktails",
  lifestyle_favAlcoolDontDrink   = "Dont Drink",
  lifestyle_hasTattoos           = "Presence de tatouages",
  lifestyle_mostFreqSocialMedia  = "Reseau social le plus utilise",
  lifestyle_socialMediaTime      = "Temps passe sur les reseaux sociaux (sur 1)",
  lifestyle_typeMorningToEvening = "Preference matinale/soiree (0 = matin, 1 = soir)",
  lifestyle_manualTaskFreq_numeric = "Fait des taches manuelles",
  dv_attitudeLeftvsRight     = "Orientation politique (0 = gauche, 1 = droite)",
  issue_proQcIndependance    = "Pro-independance du Quebec (sur 1)",
  dv_voteChoice              = "Choix de vote",
  rci_LPC = "Appui moyen au LPC (sur 1)",
  rci_CPC = "Appui moyen au CPC (sur 1)",
  rci_NDP = "Appui moyen au NDP (sur 1)",
  rci_BQ = "Appui moyen au BQ (sur 1)",
  rci_GPC = "Appui moyen au GPC (sur 1)",
  rci_PPC = "Appui moyen au PPC (sur 1)"
)

# Boucle pour calculer et imprimer la statistique pour chaque variable
for (var in names(vars_to_print)) {
  label <- vars_to_print[[var]]
  values <- cluster_data[[var]]
  
  # Pour les variables numériques avec seulement 2 valeurs uniques (0/1), on calcule le pourcentage de 1
  if (is.numeric(values) && length(unique(na.omit(values))) == 2) {
    pct <- mean(values, na.rm = TRUE) * 100
    cat(label, ": ", round(pct, 1), "%\n", sep = "")
  
  # Pour les variables numériques continues, on calcule la moyenne
  } else if (is.numeric(values)) {
    mean_val <- mean(values, na.rm = TRUE)
    cat(label, ": ", round(mean_val, 1), "\n", sep = "")
  
  # Pour les variables catégorielles, on affiche la distribution en pourcentages
  } else {
    dist <- prop.table(table(values)) * 100
    cat(label, ":\n")
    print(round(dist, 1))
  }
}


names(DataValidation)
table(DataValidation$ses_religion_bin)
table(DataValidation$lifestyle_volunteeringFreq)


# Charger les packages nécessaires
library(nnet)
library(dplyr)

# Ajustement du modèle multinomial
multinom_model <- multinom(dv_voteChoice ~ ses_genderMale + 
                                           ses_languageFrench +
                                           ses_languageOther + 
                                           ses_educBHS + 
                                           ses_educUniv + 
                                           ses_incomeLow + 
                                           ses_incomeHigh +
                                           ses_immigrant +
                                           ses_ethnicityMinority +
                                           ses_sexOrientationQueer +
                                           ses_regionPrairies +
                                           ses_regionBC +
                                           ses_regionAtlantic +
                                           ses_regionQuebec +
                                           ses_regionTerritories +
                                           ses_urban +
                                           ses_age +
                                           ses_religion_bin +
                                           lifestyle_goFishingFreq_bin +
                                           lifestyle_goHuntingFreq_bin +
                                           lifestyle_goMuseumsFreq_bin +
                                           lifestyle_motorizedActFreq_bin +
                                           lifestyle_unmotorizedActFreq_bin +
                                           lifestyle_volunteeringFreq +
                                           lifestyle_typeTransportNoCar +
                                           lifestyle_eatMeatFreq +
                                           lifestyle_consCoffeeTimHortons +
                                           lifestyle_consCoffeeMcDo +
                                           lifestyle_consCoffeeStarbucks +
                                           lifestyle_consCoffeeIndependent +
                                           lifestyle_consCoffeeNone +
                                           lifestyle_favAlcoolRedWine +
                                           lifestyle_favAlcoolWhiteWine +
                                           lifestyle_favAlcoolSpirits +
                                           lifestyle_favAlcoolBeer +
                                           lifestyle_favAlcoolCocktail +
                                           lifestyle_favAlcoolDontDrink +
                                           lifestyle_hasTattoos, data = DataValidation, maxNWts = 5000)

# Affichage du résumé du modèle
summary(multinom_model)


# Probabilités prédites
Jennifer <- data.frame(ses_genderMale = 0.48, ses_languageFrench = 0.014, ses_languageOther = 0.05, ses_educBHS = 0.17, ses_educUniv = 0.56, ses_incomeLow = 0.12, ses_incomeHigh = 0.12,
  ses_immigrant = 0.20, ses_ethnicityMinority = 0.29, ses_sexOrientationQueer = 0.15, ses_regionPrairies = 0.19, ses_regionBC = 0.16, ses_regionAtlantic = 0,
  ses_regionQuebec = 0.09, ses_regionTerritories = 0.007, ses_urban =  0.81, ses_age = 51.6, ses_religion_bin = 0.65, lifestyle_goFishingFreq_bin = 0.40, lifestyle_goHuntingFreq_bin = 0.15,
  lifestyle_goMuseumsFreq_bin = 0.84, lifestyle_motorizedActFreq_bin = 0.35, lifestyle_unmotorizedActFreq_bin = 0.71, lifestyle_volunteeringFreq = 2.7,
  lifestyle_typeTransportNoCar = 0.26, lifestyle_eatMeatFreq = 0.7, lifestyle_consCoffeeTimHortons = 0.36, lifestyle_consCoffeeMcDo = 0.08, lifestyle_consCoffeeStarbucks = 0.15,
lifestyle_consCoffeeIndependent = 0.04, lifestyle_consCoffeeNone = 0.33, lifestyle_favAlcoolRedWine = 0.25, lifestyle_favAlcoolWhiteWine = 0.16, lifestyle_favAlcoolSpirits = 0.06,
  lifestyle_favAlcoolBeer = 0.09, lifestyle_favAlcoolCocktail = 0.12, lifestyle_favAlcoolDontDrink = 0.20, lifestyle_hasTattoos = 0.25)

predict(multinom_model, newdata = Jennifer, type = "probs")

Michel <- data.frame(ses_genderMale = 0.60, ses_languageFrench = 0.97, ses_languageOther = 0, ses_educBHS = 0.17, ses_educUniv = 0.42, ses_incomeLow = 0.12, ses_incomeHigh = 0.06,
  ses_immigrant = 0.04, ses_ethnicityMinority = 0.04, ses_sexOrientationQueer = 0.09, ses_regionPrairies = 0, ses_regionBC = 0, ses_regionAtlantic = 0,
  ses_regionQuebec = 0.99, ses_regionTerritories = 0, ses_urban = 0.53, ses_age = 60, ses_religion_bin = 0.62, lifestyle_goFishingFreq_bin = 0.38, lifestyle_goHuntingFreq_bin = 0.06,
  lifestyle_goMuseumsFreq_bin = 0.76, lifestyle_motorizedActFreq_bin = 0.33, lifestyle_unmotorizedActFreq_bin = 0.70, lifestyle_volunteeringFreq = 2.1,
  lifestyle_typeTransportNoCar = 0.16, lifestyle_eatMeatFreq = 0.7, lifestyle_consCoffeeTimHortons = 0.26, lifestyle_consCoffeeMcDo = 0.09, lifestyle_consCoffeeStarbucks = 0.05,
lifestyle_consCoffeeIndependent = 0.1, lifestyle_consCoffeeNone = 0.49, lifestyle_favAlcoolRedWine = 0.21, lifestyle_favAlcoolWhiteWine = 0.25, lifestyle_favAlcoolSpirits = 0.05,
  lifestyle_favAlcoolBeer = 0.17, lifestyle_favAlcoolCocktail = 0.03, lifestyle_favAlcoolDontDrink = 0.16, lifestyle_hasTattoos = 0.22)

predict(multinom_model, newdata = Michel, type = "probs")

John <- data.frame(ses_genderMale = 0.64, ses_languageFrench = 0, ses_languageOther = 0.03, ses_educBHS = 0.24, ses_educUniv = 0.40, ses_incomeLow = 0.07, ses_incomeHigh = 0.15,
  ses_immigrant = 0.17, ses_ethnicityMinority = 0.23, ses_sexOrientationQueer = 0.05, ses_regionPrairies = 0.64, ses_regionBC = 0.36, ses_regionAtlantic = 0,
  ses_regionQuebec = 0, ses_regionTerritories = 0, ses_age = 49.5, ses_urban = 0.62, ses_religion_bin = 0.70, lifestyle_goFishingFreq_bin = 0.53, lifestyle_goHuntingFreq_bin = 0.33,
  lifestyle_goMuseumsFreq_bin = 0.85, lifestyle_motorizedActFreq_bin = 0.52, lifestyle_unmotorizedActFreq_bin = 0.78, lifestyle_volunteeringFreq = 2.5,
  lifestyle_typeTransportNoCar = 0.14, lifestyle_eatMeatFreq = 0.8, lifestyle_consCoffeeTimHortons = 0.29, lifestyle_consCoffeeMcDo = 0.18, lifestyle_consCoffeeStarbucks = 0.17,
  lifestyle_consCoffeeIndependent = 0.07, lifestyle_consCoffeeNone = 0.28, lifestyle_favAlcoolRedWine = 0.12, lifestyle_favAlcoolWhiteWine = 0.06, lifestyle_favAlcoolSpirits = 0.18,
  lifestyle_favAlcoolBeer = 0.16, lifestyle_favAlcoolCocktail = 0.08, lifestyle_favAlcoolDontDrink = 0.27, lifestyle_hasTattoos = 0.26)

predict(multinom_model, newdata = John, type = "probs")

Arjun <- data.frame(ses_genderMale = 0.50, ses_languageFrench = 0.02, ses_languageOther = 0.09, ses_educBHS = 0.25, ses_educUniv = 0.47, ses_incomeLow = 0.15, ses_incomeHigh = 0.18,
  ses_immigrant = 0.32, ses_ethnicityMinority = 0.32, ses_sexOrientationQueer = 0.06, ses_regionPrairies = 0.36, ses_regionBC = 0.18, ses_regionAtlantic = 0,
  ses_regionQuebec = 0.01, ses_regionTerritories = 0, ses_age = 47.8, ses_urban = 0.63, ses_religion_bin = 0.65, lifestyle_goFishingFreq_bin = 0.38, lifestyle_goHuntingFreq_bin = 0.15,
  lifestyle_goMuseumsFreq_bin = 0.86, lifestyle_motorizedActFreq_bin = 0.39, lifestyle_unmotorizedActFreq_bin = 0.75, lifestyle_volunteeringFreq = 2.5,
  lifestyle_typeTransportNoCar = 0.25, lifestyle_eatMeatFreq = 0.7, lifestyle_consCoffeeTimHortons = 0.31, lifestyle_consCoffeeMcDo = 0.09, lifestyle_consCoffeeStarbucks = 0.1,
  lifestyle_consCoffeeIndependent = 0.04, lifestyle_consCoffeeNone = 0.44, lifestyle_favAlcoolRedWine = 0.14, lifestyle_favAlcoolWhiteWine = 0.12, lifestyle_favAlcoolSpirits = 0.11,
  lifestyle_favAlcoolBeer = 0.09, lifestyle_favAlcoolCocktail = 0.11, lifestyle_favAlcoolDontDrink = 0.28, lifestyle_hasTattoos = 0.23)

predict(multinom_model, newdata = Arjun, type = "probs")

David <- data.frame(ses_genderMale = 0.51, ses_languageFrench = 0.02, ses_languageOther = 0.1, ses_educBHS = 0.11, ses_educUniv = 0.58, ses_incomeLow = 0.12, ses_incomeHigh = 0.27,
  ses_immigrant = 0.24, ses_ethnicityMinority = 0.27, ses_sexOrientationQueer = 0.07, ses_regionPrairies = 0, ses_regionBC = 0, ses_regionAtlantic = 0,
  ses_regionQuebec = 0.008, ses_regionTerritories = 0, ses_urban = 0.67, ses_age = 49.5, ses_religion_bin = 0.80, lifestyle_goFishingFreq_bin = 0.51, lifestyle_goHuntingFreq_bin = 0.18,
  lifestyle_goMuseumsFreq_bin = 0.78, lifestyle_motorizedActFreq_bin = 0.48, lifestyle_unmotorizedActFreq_bin = 0.77, lifestyle_volunteeringFreq = 2.5,
  lifestyle_typeTransportNoCar = 0.15, lifestyle_eatMeatFreq = 0.7, lifestyle_consCoffeeTimHortons = 0.40, lifestyle_consCoffeeMcDo = 0.12, lifestyle_consCoffeeStarbucks = 0.15,
  lifestyle_consCoffeeIndependent = 0.03, lifestyle_consCoffeeNone = 0.28, lifestyle_favAlcoolRedWine = 0.17, lifestyle_favAlcoolWhiteWine = 0.11, lifestyle_favAlcoolSpirits = 0.08,
  lifestyle_favAlcoolBeer = 0.15, lifestyle_favAlcoolCocktail = 0.06, lifestyle_favAlcoolDontDrink = 0.27, lifestyle_hasTattoos = 0.26)

predict(multinom_model, newdata = David, type = "probs")

Maxime <- data.frame(ses_genderMale = 0.68, ses_languageFrench = 0.87, ses_languageOther = 0.04, ses_educBHS = 0.23, ses_educUniv = 0.38, ses_incomeLow = 0.13, ses_incomeHigh = 0.19,
  ses_immigrant = 0.13, ses_ethnicityMinority = 0.19, ses_sexOrientationQueer = 0.10, ses_regionPrairies = 0, ses_regionBC = 0, ses_regionAtlantic = 0,
  ses_regionQuebec = 1, ses_regionTerritories = 0, ses_urban = 0.62, ses_age = 46.7, ses_religion_bin = 0.72, lifestyle_goFishingFreq_bin = 0.53, lifestyle_goHuntingFreq_bin = 0.27,
  lifestyle_goMuseumsFreq_bin = 0.77, lifestyle_motorizedActFreq_bin = 0.51, lifestyle_unmotorizedActFreq_bin = 0.81, lifestyle_volunteeringFreq = 2.3,
  lifestyle_typeTransportNoCar = 0.20, lifestyle_eatMeatFreq = 0.7, lifestyle_consCoffeeTimHortons = 0.35, lifestyle_consCoffeeMcDo = 0.1, lifestyle_consCoffeeStarbucks = 0.13,
  lifestyle_consCoffeeIndependent = 0.08, lifestyle_consCoffeeNone = 0.29, lifestyle_favAlcoolRedWine = 0.17, lifestyle_favAlcoolWhiteWine = 0.14, lifestyle_favAlcoolSpirits = 0.06,
  lifestyle_favAlcoolBeer = 0.22, lifestyle_favAlcoolCocktail = 0.11, lifestyle_favAlcoolDontDrink = 0.19, lifestyle_hasTattoos = 0.25)

predict(multinom_model, newdata = Maxime, type = "probs")

Zoe <- data.frame(ses_genderMale = 0.44, ses_languageFrench = 0.17, ses_languageOther = 0.02, ses_educBHS = 0.17, ses_educUniv = 0.51, ses_incomeLow = 0.14, ses_incomeHigh = 0.17,
  ses_immigrant = 0.17, ses_ethnicityMinority = 0.25, ses_sexOrientationQueer = 0.29, ses_regionPrairies = 0.20, ses_regionBC = 0.23, ses_regionAtlantic = 0,
  ses_regionQuebec = 0.20, ses_regionTerritories = 0.007, ses_urban = 0.70, ses_age = 44, ses_religion_bin = 0.51, lifestyle_goFishingFreq_bin = 0.41, lifestyle_goHuntingFreq_bin = 0.18,
  lifestyle_goMuseumsFreq_bin = 0.85, lifestyle_motorizedActFreq_bin = 0.38, lifestyle_unmotorizedActFreq_bin = 0.78, lifestyle_volunteeringFreq = 2.6,
  lifestyle_typeTransportNoCar = 0.32, lifestyle_eatMeatFreq = 0.6, lifestyle_consCoffeeTimHortons = 0.29, lifestyle_consCoffeeMcDo = 0.11, lifestyle_consCoffeeStarbucks = 0.16,
  lifestyle_consCoffeeIndependent = 0.11, lifestyle_consCoffeeNone = 0.30,lifestyle_favAlcoolRedWine = 0.16, lifestyle_favAlcoolWhiteWine = 0.08, lifestyle_favAlcoolSpirits = 0.09,
  lifestyle_favAlcoolBeer = 0.09, lifestyle_favAlcoolCocktail = 0.19, lifestyle_favAlcoolDontDrink = 0.25, lifestyle_hasTattoos = 0.46)

predict(multinom_model, newdata = Zoe, type = "probs")

Julie <- data.frame(ses_genderMale = 0.40, ses_languageFrench = 0.92, ses_languageOther = 0, ses_educBHS = 0.36, ses_educUniv = 0.33, ses_incomeLow = 0.26, ses_incomeHigh = 0.10,
  ses_immigrant = 0.11, ses_ethnicityMinority = 0.03, ses_sexOrientationQueer = 0.07, ses_regionPrairies = 0, ses_regionBC = 0, ses_regionAtlantic = 0,
  ses_regionQuebec = 0.99, ses_regionTerritories = 0, ses_urban = 0.51, ses_age = 53, ses_religion_bin = 0.82, lifestyle_goFishingFreq_bin = 0.28, lifestyle_goHuntingFreq_bin = 0.09,
  lifestyle_goMuseumsFreq_bin = 0.68, lifestyle_motorizedActFreq_bin = 0.39, lifestyle_unmotorizedActFreq_bin = 0.74, lifestyle_volunteeringFreq = 2,
  lifestyle_typeTransportNoCar = 0.11, lifestyle_eatMeatFreq = 0.7, lifestyle_consCoffeeTimHortons = 0.27, lifestyle_consCoffeeMcDo = 0.11, lifestyle_consCoffeeStarbucks = 0.07,
  lifestyle_consCoffeeIndependent = 0.07, lifestyle_consCoffeeNone = 0.47,lifestyle_favAlcoolRedWine = 0.22, lifestyle_favAlcoolWhiteWine = 0.18, lifestyle_favAlcoolSpirits = 0.05,
  lifestyle_favAlcoolBeer = 0.15, lifestyle_favAlcoolCocktail = 0.10, lifestyle_favAlcoolDontDrink = 0.21, lifestyle_hasTattoos = 0.41)

predict(multinom_model, newdata = Julie, type = "probs")

Robert <- data.frame(ses_genderMale = 0.60, ses_languageFrench = 0, ses_languageOther = 0, ses_educBHS = 0.30, ses_educUniv = 0.30, ses_incomeLow = 0.14, ses_incomeHigh = 0.05,
  ses_immigrant = 0.19, ses_ethnicityMinority = 0.05, ses_sexOrientationQueer = 0.22, ses_regionPrairies = 0, ses_regionBC = 0, ses_regionAtlantic = 1,
  ses_regionQuebec = 0, ses_regionTerritories = 0, ses_urban = 0.35, ses_age = 54, ses_religion_bin = 0.76, lifestyle_goFishingFreq_bin = 0.60, lifestyle_goHuntingFreq_bin = 0.27,
  lifestyle_goMuseumsFreq_bin = 0.68, lifestyle_motorizedActFreq_bin = 0.38, lifestyle_unmotorizedActFreq_bin = 0.76, lifestyle_volunteeringFreq = 2.8,
  lifestyle_typeTransportNoCar = 0.16, lifestyle_eatMeatFreq = 0.7, lifestyle_consCoffeeTimHortons = 0.35, lifestyle_consCoffeeMcDo = 0.03, lifestyle_consCoffeeStarbucks = 0.11,
  lifestyle_consCoffeeIndependent = 0.05, lifestyle_consCoffeeNone = 0.43,lifestyle_favAlcoolRedWine = 0.05, lifestyle_favAlcoolWhiteWine = 0.05, lifestyle_favAlcoolSpirits = 0.19,
  lifestyle_favAlcoolBeer = 0.16, lifestyle_favAlcoolCocktail = 0.11, lifestyle_favAlcoolDontDrink = 0.24, lifestyle_hasTattoos = 0.27)

predict(multinom_model, newdata = Robert, type = "probs")

Emily <- data.frame(ses_genderMale = 0.54, ses_languageFrench = 0.19, ses_languageOther = 0.08, ses_educBHS = 0.27, ses_educUniv = 0.54, ses_incomeLow = 0.22, ses_incomeHigh = 0.08,
  ses_immigrant = 0.27, ses_ethnicityMinority = 0.16, ses_sexOrientationQueer = 0.27, ses_regionPrairies = 0, ses_regionBC = 0.38, ses_regionAtlantic = 0.03,
  ses_regionQuebec = 0.27, ses_regionTerritories = 0.03, ses_urban = 0.60, ses_age = 45, ses_religion_bin = 0.46, lifestyle_goFishingFreq_bin = 0.27, lifestyle_goHuntingFreq_bin = 0.14,
  lifestyle_goMuseumsFreq_bin = 0.81, lifestyle_motorizedActFreq_bin = 0.35, lifestyle_unmotorizedActFreq_bin = 0.78, lifestyle_volunteeringFreq = 2.5,
  lifestyle_typeTransportNoCar = 0.19, lifestyle_eatMeatFreq = 0.7, lifestyle_consCoffeeTimHortons = 0.27, lifestyle_consCoffeeMcDo = 0, lifestyle_consCoffeeStarbucks = 0.11,
  lifestyle_consCoffeeIndependent = 0.16, lifestyle_consCoffeeNone = 0.46,lifestyle_favAlcoolRedWine = 0.14, lifestyle_favAlcoolWhiteWine = 0.08, lifestyle_favAlcoolSpirits = 0.08,
  lifestyle_favAlcoolBeer = 0.19, lifestyle_favAlcoolCocktail = 0.22, lifestyle_favAlcoolDontDrink = 0.16, lifestyle_hasTattoos = 0.32)

predict(multinom_model, newdata = Emily, type = "probs")
